requirements.txt
celery[redis]==5.3.4
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
redis==5.0.1
